backend/ai-tutor-python/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
openai>=1.30.1
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
authlib>=1.3.1
itsdangerous>=2.2.0
langgraph>=0.2.0
langchain-openai>=0.1.20
langchain-core>=0.2.40
